Localization works properly for header posts and archive
Basically it seems that paginate-v2 and polyglot cannot work together.
The solution is to disable polyglot and use paginate-v2 to handle the
localization of the archive posts.
Instead, the posts displayed at the top of the page must be handled by using two
different index.html files, where the main file contains all all the necessary
code, while the index_it.html is only used to manage the localization.
For some reason (bug?), this latter file must contain at least one liquid command
to work properly, which is not used.